Extract out profile container
