+++ pages/profiles/[wallet].tsx
@@ -1,19 +1,11 @@
 import { ActivityContent } from '@/common/components/elements/ActivityContent';
-import Image from 'next/image';
 import Head from 'next/head';
 import { GetServerSideProps } from 'next';
-import { WalletPill } from '@/common/components/elements/WalletIndicator';
-import { useEffect, useState } from 'react';
-import styled from 'styled-components';
-import { useWalletProfileLazyQuery } from 'src/graphql/indexerTypes';
-import { useTwitterHandle } from '@/common/hooks/useTwitterHandle';
 import { PublicKey } from '@solana/web3.js';
-import { mq } from '@/common/styles/MediaQuery';
 import { showFirstAndLastFour } from '@/modules/utils/string';
-import Bugsnag from '@bugsnag/js';
+import { ProfileContainer } from '@/common/components/elements/ProfileContainer';
 
 export const getServerSideProps: GetServerSideProps = async (context) => {
-  // ...
   return {
     props: {
       // query params must be gotten serverside to be available on initial render
@@ -24,56 +16,6 @@ export const getServerSideProps: GetServerSideProps = async (context) => {
 
 const ActivityLanding = ({ wallet }: { wallet: string }) => {
   const publicKey = wallet ? new PublicKey(wallet as string) : null;
-  const [queryWalletProfile, walletProfile] = useWalletProfileLazyQuery();
-  const { data: twitterHandle } = useTwitterHandle(publicKey);
-
-  const [{ pfp, banner }, setPfpAndBanner] = useState({
-    pfp: '/images/gradients/gradient-3.png',
-    banner: 'url(/images/gradients/gradient-5.png)', // TODO: Fetch from wallet (DERIVE)
-  });
-
-  useEffect(() => {
-    if (!twitterHandle) return;
-    try {
-      queryWalletProfile({
-        variables: {
-          handle: twitterHandle,
-        },
-      });
-    } catch (error: any) {
-      console.error(error);
-      console.log('failed to fetch wallet');
-      Bugsnag.notify(error);
-    }
-  }, [queryWalletProfile, twitterHandle]);
-
-  const bannerUrl = walletProfile.data?.profile?.bannerImageUrl;
-  const imageUrl = walletProfile.data?.profile?.profileImageUrlHighres?.replace('_normal', '');
-
-  const bannerBackgroundImage = !!bannerUrl
-    ? `url(${bannerUrl})`
-    : 'url(/images/gradients/gradient-5.png)'; // TODO: Fetch from wallet (DERIVE).
-  const profilePictureImage = imageUrl ?? '/images/gradients/gradient-3.png'; // TODO: Fetch from wallet [here-too] (DERIVE).
-
-  useEffect(() => {
-    const profilePictureImage = imageUrl ?? '/images/gradients/gradient-3.png'; // TODO: Fetch from wallet [here-too] (DERIVE).
-    const bannerBackgroundImage = !!bannerUrl
-      ? `url(${bannerUrl})`
-      : 'url(/images/gradients/gradient-5.png)'; // TODO: Fetch from wallet (DERIVE).
-
-    setPfpAndBanner({
-      pfp: profilePictureImage,
-      banner: bannerBackgroundImage,
-    });
-  }, [imageUrl, bannerUrl]);
-
-  const getPublicKeyFromWalletOnUrl = () => {
-    try {
-      return new PublicKey(wallet as string);
-    } catch (_) {
-      return null;
-    }
-  };
 
   return (
     <>
@@ -85,92 +27,12 @@ const ActivityLanding = ({ wallet }: { wallet: string }) => {
           content="View activity for this, or any other pubkey, in the Holaplex ecosystem."
         />
       </Head>
-      <HeadingContainer>
-        <Banner style={{ backgroundImage: banner }} />
-      </HeadingContainer>
-      <ContentCol>
-        <Profile>
-          <ProfilePictureContainer>
-            <ProfilePicture src={pfp} width={PFP_SIZE} height={PFP_SIZE} />
-          </ProfilePictureContainer>
-          <WalletPillContainer>
-            <WalletPill
-              disableBackground
-              disableLink
-              textOverride={twitterHandle ? `${twitterHandle}` : null}
-              publicKey={getPublicKeyFromWalletOnUrl()}
-            />
-          </WalletPillContainer>
-        </Profile>
-        <ActivityContentWrapper>
-          <ActivityContent publicKey={publicKey} />
-        </ActivityContentWrapper>
-      </ContentCol>
+
+      <ProfileContainer wallet={wallet} publicKey={publicKey}>
+        <ActivityContent publicKey={publicKey} />
+      </ProfileContainer>
     </>
   );
 };
 
 export default ActivityLanding;
-
-const PFP_SIZE = 90;
-const BOX_SIZE = 1400;
-
-const WalletPillContainer = styled.div`
-  margin-top: 80px;
-`;
-
-const Profile = styled.div`
-  min-width: 348px;
-  position: relative;
-`;
-
-const ActivityContentWrapper = styled.section`
-  margin-top: ${PFP_SIZE / 2}px;
-  width: 100%;
-`;
-
-const ContentCol = styled.div`
-  width: 100%;
-  margin-left: auto;
-  margin-right: auto;
-  display: flex;
-  flex-direction: column;
-  padding-left: 20px;
-  padding-right: 20px;
-  ${mq('md')} {
-    padding-left: ${PFP_SIZE - 40}px;
-    padding-right: ${PFP_SIZE - 40}px;
-    max-width: ${BOX_SIZE}px;
-    flex-direction: row;
-  }
-  ${mq('lg')} {
-    padding-left: ${PFP_SIZE - 20}px;
-    padding-right: ${PFP_SIZE - 20}px;
-  }
-`;
-
-const HeadingContainer = styled.header``;
-
-const ProfilePictureContainer = styled.div`
-  position: absolute;
-  top: ${-PFP_SIZE / 2}px;
-  @media (min-width: ${BOX_SIZE - PFP_SIZE}) {
-    left: 0px;
-  }
-`;
-
-const ProfilePicture = styled(Image)`
-  border-radius: 50%;
-  border: 5px solid #161616 !important;
-`;
-
-const Banner = styled.div`
-  width: 100%;
-  height: 265px;
-  background-repeat: no-repeat;
-  background-size: cover;
-  ${mq('lg')} {
-    background-attachment: fixed;
-    background-size: 100%;
-  }
-`;
